Home user component finalized
