+++ Web Application/web_application_frontend/src/App.js
@@ -1,6 +1,7 @@
 import React from "react";
 import NavBar from "./components/Navbar";
-import Home from "../src/components/Users Home/Home_User"
+import Home from "../src/components/Users Home/Home_User_New"
+
 
 function App() {
   return (
