--- Web Application/web_application_frontend/src/components/Users Home/Home_User.css
@@ -1,23 +1,0 @@
-.parent-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-
-.plant-container{
-    width: 600px;
-    height: 150px;
-}
-
-.row {
-    display: flex;
-    align-items: center;
-    margin-bottom: 20px;
-    box-shadow: 15px 15px 20px rgba(0, 0, 0, 0.3);
-  }
-
-  
-.plant-card-items{
-    background-color: rgb(183, 232, 167);
-}
